requirements.txt
uvicorn==0.17.6
fastapi==0.75.2
pydantic==1.9.0
SQLAlchemy==1.4.36
psycopg2==2.7.7
python-dotenv==0.20.0
alembic==1.7.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5